ESP-IDF: fix the 29 remaining example/test build failures (#186)
All 225 ESP-IDF apps (171 examples + 54 tests) now build on v6.1.

Drivers (shared by every C++ platform):
- BMP085/BMP180: the "0xFFFF = no device" calibration check compared
  the int16_t coefficients (AC1-3, B1, B2, MB, MC, MD) with 0xFFFF,
  which is never true, so an all-ones read went undetected for them.
  Compare as uint16_t.
- PCF8574/PCF8575: initializer list order now matches member
  declaration order (-Wreorder).

Examples:
- L3GD20H (all 3): used a nonexistent I2CConnectionESPIDF(port, addr)
  constructor; now sets up the bus like every other ESP-IDF I2C app.
  complete no longer calls who_am_i(), which isn't in the API (the
  constructor checks WHO_AM_I).
- BME680 complete/demo: FILTER_OFF -> FILTER_0.
- BMP280 complete: assign `alt` before using it.
- PCF8574 minimal: write_port() takes (port, mask).
- MCP2515 demo: cast the CAN id for %X (uint32_t is long on Xtensa).

Tests:
- l3gd20h: same bus setup fix; rewritten to the standard
  PASS/FAIL + ===DONE=== format so test_espidf.sh can parse it.
- pcf8591, l3g4200d, mpu9250, apds9960: removed checks that could
  never fail (e.g. uint8_t <= 255) and tripped -Wtype-limits.
- bmp581: dropped the check that read protected members; fixed
  BMP581Full.MODE_NORMAL -> BMP581Full::MODE_NORMAL.

+++ cpp/examples/espidf/comms/MCP2515/demo/main/main.cpp
@@ -56,7 +56,7 @@ extern "C" void app_main(void) {
         bool got = mcp2515.recv(frame, 100);                          // Poll for a received frame, (frame, timeout_ms=100) → bool
 
         if (got) {
-            printf("RX id=0x%X dlc=%u data=", frame.extended ? frame.id : (frame.id & 0x7FF),
+            printf("RX id=0x%X dlc=%u data=", (unsigned)(frame.extended ? frame.id : (frame.id & 0x7FF)),
                    frame.dlc);
             for (uint8_t i = 0; i < frame.dlc; i++) printf("%02X ", frame.data[i]);
             printf("[%s%s]\n", frame.extended ? "EXT" : "STD", frame.rtr ? "/RTR" : "");
